Add user setting for redirectFromHome
