+++ lib/controllers/middleware.js
@@ -1,8 +1,15 @@
 const errors = require('../errors');
+const config = require('../config');
 const rateLimit = require('../services/rate-limiting');
 
 const authRateLimit = rateLimit('unauthenticated');
 const noAuthRateLimit = rateLimit('authenticated');
+
+const COOKIE_OPTIONS = {
+	httpOnly: true,
+	secure: 'auto',
+	domain: config.get('domain')
+};
 
 function requireConsent(req, res, next) {
 	if (req.session.userProfile) {
@@ -18,6 +25,19 @@ function requireConsent(req, res, next) {
 }
 
 module.exports = {
+	addRedirectCookie(req, res, next) {
+		const userSettings = JSON.parse(req.user.settings);
+
+		if (userSettings && userSettings.redirectFromHome) {
+			// Gradebook auto redirect
+			// @todo: compute domain slug
+			res.cookie('gbardr', 'dinklemeyer', COOKIE_OPTIONS);
+		} else {
+			res.clearCookie('gbardr');
+		}
+
+		next();
+	},
 	requireAuth(req, res, next) {
 		if (req.user) {
 			return next();
