+++ lib/controllers/user.js
@@ -3,6 +3,7 @@ const {user: {response: UserModel}} = require('../models');
 const api = require('../api');
 const themeService = require('../services/theming');
 const {user: sanitize} = require('./sanitizers');
+const {addRedirectCookie} = require('./middleware');
 
 module.exports = {
 	me: (req, res) => {
@@ -76,7 +77,7 @@ module.exports = {
 
 		await user.commit();
 
-		res.status(200).set('content-type', 'application/json').end(settingsString);
+		addRedirectCookie(req, res, () => res.status(200).set('content-type', 'application/json').end(settingsString));
 	},
 	async delete(req, res) {
 		if (req.session.userProfile) {
