+++ lib/services/validation/user-settings.js
@@ -1,10 +1,13 @@
 const errors = require('../../errors');
 
+const isBoolean = value => typeof value === 'boolean';
+
 const SETTINGS = {
 	tour: {
-		isValid(value) {
-			return typeof value === 'boolean';
-		}
+		isValid: isBoolean
+	},
+	redirectFromHome: {
+		isValid: isBoolean
 	}
 };
